main,refactor: split kind.h into kind.h and kind_p.h
