+++ main/writer-json.c
@@ -14,6 +14,7 @@
 #include "mio.h"
 #include "options_p.h"
 #include "read.h"
+#include "routines.h"
 #include "ptag_p.h"
 #include "writer_p.h"
