+++ win32/ctags_vs2013.vcxproj.filters
@@ -560,6 +560,9 @@
     <ClInclude Include="..\main\kind.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\main\kind_p.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="..\main\lregex.h">
       <Filter>Header Files</Filter>
     </ClInclude>
